+++ common/src/main/java/mod/adrenix/nostalgic/common/config/DefaultConfig.java
@@ -164,6 +164,7 @@ public abstract class DefaultConfig
         public static final boolean OLD_BACKWARD_WALKING = true;
         public static final boolean OLD_COLLIDE_BOBBING = true;
         public static final boolean OLD_VERTICAL_BOBBING = true;
+        public static final boolean OLD_CREATIVE_CROUCH = true;
         public static final boolean OLD_SNEAKING = true;
     }