+++ common/src/main/java/mod/adrenix/nostalgic/mixin/client/LocalPlayerMixin.java
@@ -8,9 +8,11 @@ import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.player.AbstractClientPlayer;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.world.InteractionHand;
+import net.minecraft.world.entity.Pose;
 import net.minecraft.world.entity.player.ProfilePublicKey;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -26,6 +28,11 @@ public abstract class LocalPlayerMixin extends AbstractClientPlayer implements I
     {
         super(level, profile, key);
     }
+
+    /* Shadows */
+
+    @Shadow private boolean crouching;
+    @Shadow public abstract boolean isShiftKeyDown();
 
     /* Client Side Swing Ducking */
 
@@ -70,5 +77,16 @@ public abstract class LocalPlayerMixin extends AbstractClientPlayer implements I
 
         if (ModConfig.Animation.oldSwingDropping() && NostalgicTweaks.isNetworkVerified() && isSwingBlocked)
             callback.cancel();
+    }
+
+    /**
+     * Lets the crouching state be active in while the player is in creative mode.
+     * Controlled by the old creative crouch animation tweak.
+     */
+    @Inject(method = "aiStep", at = @At(value = "INVOKE", target = "Lnet/minecraft/util/Mth;clamp(FFF)F"))
+    private void NT$onAiStep(CallbackInfo callback)
+    {
+        if (ModConfig.Animation.oldCreativeCrouch() && this.getAbilities().flying)
+            this.crouching = !this.isSwimming() && this.canEnterPose(Pose.CROUCHING) && (this.isShiftKeyDown() || !this.isSleeping() && !this.canEnterPose(Pose.STANDING));
     }
 }