+++ common/src/main/java/mod/adrenix/nostalgic/mixin/common/world/entity/PlayerMixin.java
@@ -1,6 +1,7 @@
 package mod.adrenix.nostalgic.mixin.common.world.entity;
 
 import com.mojang.authlib.GameProfile;
+import mod.adrenix.nostalgic.NostalgicTweaks;
 import mod.adrenix.nostalgic.common.config.ModConfig;
 import mod.adrenix.nostalgic.common.gameplay.OldFoodData;
 import mod.adrenix.nostalgic.mixin.widen.IMixinLivingEntity;
@@ -10,6 +11,7 @@ import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Pose;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.player.ProfilePublicKey;
 import net.minecraft.world.food.FoodData;
@@ -19,6 +21,7 @@ import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.ModifyArg;
 import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
@@ -99,5 +102,28 @@ public abstract class PlayerMixin extends LivingEntity
         if (ModConfig.Gameplay.disableSweep())
             return new ArrayList<>();
         return instance.getEntitiesOfClass(aClass, aabb);
+    }
+
+    /**
+     * Changes the connected player's crouching pose when flying in creative.
+     * Controlled by the old creative crouch animation tweak.
+     */
+    @ModifyArg(method = "updatePlayerPose", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/player/Player;setPose(Lnet/minecraft/world/entity/Pose;)V"))
+    private Pose NT$onSetPlayerPose(Pose vanilla)
+    {
+        if (!ModConfig.Animation.oldCreativeCrouch() || !NostalgicTweaks.isNetworkVerified())
+            return vanilla;
+
+        Pose pose = this.isFallFlying() ? Pose.FALL_FLYING :
+            this.isSleeping() ? Pose.SLEEPING :
+            this.isSwimming() ? Pose.SWIMMING :
+            this.isAutoSpinAttack() ? Pose.SPIN_ATTACK :
+            this.isShiftKeyDown() ? Pose.CROUCHING :
+            Pose.STANDING
+        ;
+
+        return this.isSpectator() || this.isPassenger() || this.canEnterPose(pose) ? pose :
+            this.canEnterPose(Pose.CROUCHING) ? Pose.CROUCHING : Pose.SWIMMING
+        ;
     }
 }