+++ changelog.md
@@ -49,4 +49,5 @@ All updates that start with a 2.x.x version number will support multiplayer.
 - Added new tweak: (Hunger System) Old Food Stacking
 > Only vanilla items will be impacted. Foods that restore half a heart can stack up to 8 like potatoes and carrots.
 - Added new tweak: (Player Animation) Old Backwards Walking
+- Added new tweak: (Player Animation) Old Creative Crouching
 - Fixed the old fog rendering from overriding the dark fog caused by the darkness effect.